Create `createTransitionObj` function
