+++ src/extensions/styles/transitions/transitionAttributesCreator.js
@@ -1,4 +1,4 @@
-import breakpointAttributesCreator from '../breakpointAttributesCreator';
+import createTransitionObj from './createTransitionObj';
 import transitionDefault from './transitionDefault';
 
 const getHoverProp = (property, prefix = '') => {
@@ -15,30 +15,6 @@ const getHoverProp = (property, prefix = '') => {
 			}-status-hover`;
 	}
 };
-
-// AFTER #3265 WILL BE RECEIVED FROM THE STORE
-const disableDefaultTransition = false;
-
-const transitionRaw = breakpointAttributesCreator({
-	obj: {
-		'transition-duration': {
-			type: 'number',
-			default: 0.3,
-		},
-		'transition-delay': {
-			type: 'number',
-			default: 0,
-		},
-		easing: {
-			type: 'string',
-			default: 'ease',
-		},
-		'transition-status': {
-			type: 'boolean',
-			default: !disableDefaultTransition,
-		},
-	},
-});
 
 const transitionAttributesCreator = (transitionObj = transitionDefault) => {
 	const getTransitionOptions = transitionObj =>
@@ -57,10 +33,7 @@ const transitionAttributesCreator = (transitionObj = transitionDefault) => {
 	const blockOptions = getTransitionOptions(transitionObj?.block);
 	const canvasOptions = getTransitionOptions(transitionObj?.canvas);
 
-	const transitionRawObj = Object.keys(transitionRaw).reduce((acc, key) => {
-		acc[key] = transitionRaw[key].default;
-		return acc;
-	}, {});
+	const transitionRawObj = createTransitionObj();
 
 	const transitionStyleObj = {
 		block: {},
